Two new tests for move only types
