+++ test/future_when_all_arguments_tests.cpp
@@ -12,10 +12,10 @@
 
 #include <string>
 
-#include "test_helper.hpp"
+#include "future_test_helper.hpp"
 
 using namespace stlab;
-using namespace test_helper;
+using namespace future_test_helper;
 
 BOOST_FIXTURE_TEST_SUITE(future_when_all_args_int, test_fixture<int>)
     BOOST_AUTO_TEST_CASE(future_when_all_args_int_with_one_element) {
